Changed type popularity Int to Double in the class Movie.kt; Created ambient Retrofit and repository; Created viewModel DetailMovieViewModel.kt and collect info for Activity; Created request of the viewModel and created observer for change views in the DetailMovieActivity.kt.

+++ app/src/main/java/br/com/jlcampos/movie2you/data/model/Movie.kt
@@ -17,7 +17,7 @@ data class Movie (
     @SerializedName(Constants.movie_vote_count)
     var voteCount: Int,
     @SerializedName(Constants.movie_popularity)
-    var popularity: Int,
+    var popularity: Double,
     @SerializedName(Constants.movie_release_date)
     var releaseDate: String?
 )
